+++ client/Packages/com.beamable.server/Editor/UI/Model/BeamoServiceBuilder.cs
@@ -7,6 +7,7 @@ namespace Beamable.Editor.UI.Model
 {
 	public class BeamoServiceBuilder : IBeamableBuilder
 	{
+
 		public Task CheckIfIsRunning()
 		{
 			return Task.CompletedTask;
@@ -33,7 +34,18 @@ namespace Beamable.Editor.UI.Model
 		public Action<int, int> OnStartingProgress { get; set; }
 		public Action<bool> OnStartingFinished { get; set; }
 		public Action<bool> OnBuildingFinished { get; set; }
-		public bool IsRunning { get; set; }
+		public bool IsRunning
+		{
+			get => _isRunning;
+			set
+			{
+				if (value == _isRunning) return;
+				_isRunning = value;
+				BeamEditorContext.Default.Dispatcher.Schedule(() => OnIsRunningChanged?.Invoke(value));
+			}
+		}
+
 		private CodeService CodeService => BeamEditorContext.Default.ServiceScope.GetService<CodeService>();
+		private bool _isRunning;
 	}
 }