+++ client/Packages/com.beamable.server/Editor/UI2/Models/MicroserviceVisualsModel.cs
@@ -1,7 +1,10 @@
-﻿using Beamable.Editor.BeamCli.Commands;
+﻿using Beamable.Common.BeamCli;
+using Beamable.Editor.BeamCli.Commands;
 using Beamable.Editor.UI.Model;
 using Beamable.Server.Editor.Usam;
 using System;
+using System.IO;
+using System.Linq;
 using UnityEditor;
 using UnityEngine;
 using UnityEngine.UIElements;
@@ -20,12 +23,14 @@ namespace Beamable.Editor.Microservice.UI2.Models
 		Action<bool> OnLogsAttachmentChanged { get; set; }
 		void PopulateMoreDropdown(ContextualMenuPopulateEvent evt);
 	}
+
 	public interface IMicroserviceVisualsModel : IServiceLogsVisualModel
 	{
 		float ElementHeight { get; set; }
 		bool IsSelected { get; set; }
 		bool IsCollapsed { get; set; }
 
+		string Name { get; }
 		Action<bool> OnSelectionChanged { get; set; }
 		Action OnSortChanged { get; set; }
 	}
@@ -35,13 +40,20 @@ namespace Beamable.Editor.Microservice.UI2.Models
 	{
 		private const float DEFAULT_HEIGHT = 300.0f;
 
-		public bool AreLogsAttached { get; set; }
+		public bool AreLogsAttached
+		{
+			get => _areLogsAttached;
+			protected set => _areLogsAttached = value;
+		}
+
+		public string Name => _name;
 
 		public float ElementHeight
 		{
 			get => _visualHeight;
 			set => _visualHeight = value;
 		}
+
 		public bool IsSelected
 		{
 			get => _isSelected;
@@ -51,6 +63,7 @@ namespace Beamable.Editor.Microservice.UI2.Models
 				OnSelectionChanged?.Invoke(value);
 			}
 		}
+
 		public bool IsCollapsed
 		{
 			get => _isCollapsed;
@@ -68,40 +81,29 @@ namespace Beamable.Editor.Microservice.UI2.Models
 		[SerializeField] private bool _isSelected;
 		[SerializeField] private bool _isCollapsed = false;
 		[SerializeField] private float _visualHeight = DEFAULT_HEIGHT;
-		[SerializeField] public string name;
+		[SerializeField] public string _name;
+		[SerializeField] private bool _areLogsAttached = true;
+		private IBeamoServiceDefinition _serviceDefinition;
 
-		public static MicroserviceVisualsModel GetModel(string name)
+		private void HandleLogMessage(string serviceName, BeamTailLogMessage log)
 		{
-			MicroserviceVisualsModel model = null;
-			try
+			if (serviceName == _name && !string.IsNullOrWhiteSpace(log.message))
 			{
-				var key = GetKey(name);
-				if (EditorPrefs.HasKey(key))
-				{
-					string json = EditorPrefs.GetString(key, string.Empty);
-					model = JsonUtility.FromJson<MicroserviceVisualsModel>(json);
-				}
-				if (model == null)
-				{
-					model = new MicroserviceVisualsModel() { name = name };
-				}
-
-				BeamEditorContext.Default.ServiceScope.GetService<CodeService>().OnLogMessage -= model.HandleLogMessage;
-				BeamEditorContext.Default.ServiceScope.GetService<CodeService>().OnLogMessage += model.HandleLogMessage;
+				Logs.AddMessage(FromBeamTailLog(log));
 			}
-			catch
-			{
-				//
-			}
-			return model;
 		}
 
-		private void HandleLogMessage(string arg1, BeamTailLogMessage arg2)
+		public void Disconnect()
 		{
-			if (arg1 == name && !string.IsNullOrWhiteSpace(arg2.message))
-			{
-				Logs.AddMessage(FromBeamTailLog(arg2));
-			}
+			BeamEditorContext.Default.ServiceScope.GetService<CodeService>().OnLogMessage -= HandleLogMessage;
+		}
+
+		public void Connect()
+		{
+			_serviceDefinition = BeamEditorContext.Default.ServiceScope.GetService<CodeService>().ServiceDefinitions
+							 .FirstOrDefault(def => def.ServiceInfo.name == Name);
+			BeamEditorContext.Default.ServiceScope.GetService<CodeService>().OnLogMessage -= HandleLogMessage;
+			BeamEditorContext.Default.ServiceScope.GetService<CodeService>().OnLogMessage += HandleLogMessage;
 		}
 
 		static LogMessage FromBeamTailLog(BeamTailLogMessage message)
@@ -129,12 +131,6 @@ namespace Beamable.Editor.Microservice.UI2.Models
 			return new LogMessage() { Message = message.message, Level = logLevel, Timestamp = message.timeStamp };
 		}
 
-		public void Save()
-		{
-			string json = JsonUtility.ToJson(this);
-			EditorPrefs.SetString(GetKey(name), json);
-		}
-
 		public void DetachLogs()
 		{
 			if (!AreLogsAttached) return;
@@ -143,6 +139,7 @@ namespace Beamable.Editor.Microservice.UI2.Models
 			OnLogsDetached?.Invoke();
 			OnLogsAttachmentChanged?.Invoke(false);
 		}
+
 		public void AttachLogs()
 		{
 			if (AreLogsAttached) return;
@@ -153,8 +150,57 @@ namespace Beamable.Editor.Microservice.UI2.Models
 
 		public virtual void PopulateMoreDropdown(ContextualMenuPopulateEvent evt)
 		{
+
+			var existsOnRemote = _serviceDefinition.IsRunningOnRemote == BeamoServiceStatus.Running;
+			var imageSuffix = !string.IsNullOrWhiteSpace(_serviceDefinition.ImageId) ? string.Empty : " (Build first)";
+			var localCategory = _serviceDefinition.IsRunningLocally ? "Local" : "Local (not running)";
+			var remoteCategory = existsOnRemote ? "Cloud" : "Cloud (not deployed)";
+			evt.menu.BeamableAppendAction($"Reveal build directory{imageSuffix}", pos =>
+			{
+				var full = Path.GetFullPath(_serviceDefinition.ServiceInfo.dockerBuildPath);
+				EditorUtility.RevealInFinder(full);
+			});
+			evt.menu.BeamableAppendAction($"{localCategory}/View Documentation", pos => OpenDocs(false), _serviceDefinition.IsRunningLocally);
+			if (_serviceDefinition.ServiceInfo != null)
+			{
+				evt.menu.BeamableAppendAction($"{localCategory}/Regenerate {_serviceDefinition.BeamoId}Client.cs", pos =>
+				{
+					BeamEditorContext.Default.ServiceScope.GetService<CodeService>().GenerateClientCode(_name)
+									 .Then(_ => { });
+				});
+			}
+
+			evt.menu.BeamableAppendAction($"{remoteCategory}/View Documentation", pos => { OpenOnRemote("docs/"); }, existsOnRemote);
+			evt.menu.BeamableAppendAction($"{remoteCategory}/View Metrics", pos => { OpenOnRemote("metrics"); }, existsOnRemote);
+			evt.menu.BeamableAppendAction($"{remoteCategory}/View Logs", pos => { OpenDocs(true); }, existsOnRemote);
+			evt.menu.BeamableAppendAction($"Open C# Code", _ => _serviceDefinition.ServiceInfo.OpenCode());
+
+			if (!AreLogsAttached)
+			{
+				evt.menu.BeamableAppendAction($"Reattach Logs", pos => AttachLogs());
+			}
 		}
 
-		private static string GetKey(string name) => $"Beamable{nameof(MicroserviceVisualsModel)}{name}";
+
+		protected void OpenRemoteDocs() => OpenOnRemote("docs");
+		protected void OpenRemoteMetrics() => OpenOnRemote("metrics");
+		protected void OpenOnRemote(string relativePath)
+		{
+			var api = BeamEditorContext.Default;
+			var path =
+				$"{BeamableEnvironment.PortalUrl}/{api.CurrentCustomer.Alias}/" +
+				$"games/{api.ProductionRealm.Pid}/realms/{api.CurrentRealm.Pid}/" +
+				$"microservices/{_serviceDefinition.BeamoId}/{relativePath}?refresh_token={api.Requester.Token.RefreshToken}";
+			Application.OpenURL(path);
+		}
+
+		private void OpenDocs(bool remote)
+		{
+			if (_serviceDefinition.IsRunningLocally)
+			{
+				BeamEditorContext.Default.ServiceScope.GetService<CodeService>()
+								 .OpenSwagger(_serviceDefinition.BeamoId, remote).Then(_ => { });
+			}
+		}
 	}
 }