+++ cli/cli/CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `--dotnet-path` option is available to override dotnet installation location.
 - Auto generation handles trials and session services.
 
+### Fixed
+
+- `project add` Dockerfile path fixes.
+
 ## [1.19.7]
 
 no changes